Move doublekeydictionary to share project
