+++ Source/HelixToolkit.Shared/HelixToolkit.Shared.projitems
@@ -19,6 +19,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Geometry\SweepLinePolygonTriangulator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HelixToolkitException.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helper\ContourHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Helper\DoubleKeyDictionary.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helper\MathHelper.cs" />
   </ItemGroup>
 </Project>
